Split out null problem in its own files, some cleanups.
