+++ src/problems/unconstrain.cpp
@@ -39,6 +39,7 @@ see https://www.gnu.org/licenses/. */
 #include <pagmo/exceptions.hpp>
 #include <pagmo/io.hpp>
 #include <pagmo/problem.hpp>
+#include <pagmo/problems/null_problem.hpp>
 #include <pagmo/problems/unconstrain.hpp>
 #include <pagmo/s11n.hpp>
 #include <pagmo/threading.hpp>
