+++ tests/cec2014.cpp
@@ -136,7 +136,7 @@ BOOST_AUTO_TEST_CASE(cec2014_serialization_test)
         oarchive << p;
     }
     // Change the content of p before deserializing.
-    p = problem{null_problem{}};
+    p = problem{};
     {
         boost::archive::binary_iarchive iarchive(ss);
         iarchive >> p;
