+++ tests/golomb_ruler.cpp
@@ -109,7 +109,7 @@ BOOST_AUTO_TEST_CASE(golomb_ruler_serialization_test)
         oarchive << p;
     }
     // Change the content of p before deserializing.
-    p = problem{null_problem{}};
+    p = problem{};
     {
         boost::archive::binary_iarchive iarchive(ss);
         iarchive >> p;
