+++ tests/hock_schittkowsky_71.cpp
@@ -95,7 +95,7 @@ BOOST_AUTO_TEST_CASE(hock_schittkowsky_71_serialization_test)
         oarchive << p;
     }
     // Change the content of p before deserializing.
-    p = problem{null_problem{}};
+    p = problem{};
     {
         boost::archive::binary_iarchive iarchive(ss);
         iarchive >> p;
