+++ tests/inventory.cpp
@@ -80,7 +80,7 @@ BOOST_AUTO_TEST_CASE(inventory_serialization_test)
         oarchive << p;
     }
     // Change the content of p before deserializing.
-    p = problem{null_problem{}};
+    p = problem{};
     {
         boost::archive::binary_iarchive iarchive(ss);
         iarchive >> p;
