+++ tests/rosenbrock.cpp
@@ -94,7 +94,7 @@ BOOST_AUTO_TEST_CASE(rosenbrock_serialization_test)
         oarchive << p;
     }
     // Change the content of p before deserializing.
-    p = problem{null_problem{}};
+    p = problem{};
     {
         boost::archive::binary_iarchive iarchive(ss);
         iarchive >> p;
